SCF/requirements.txt
customtkinter>=5.2.0
pandas>=2.0.0
fpdf2>=2.7.5